Fixed magiclink problems.
Phone number regex was checking for a '+' at the start, which US
users aren't familiar with. Changed it to just check for 10 digits.

Auth service now strips dashes from emailPhone, so users can enter
dashes without it breaking. It prepends '+1' to to the number so
that SNS gets the right format.

Fixed non-rendering description and label on magiclink modal.


Former-commit-id: 74a3fa5297d9359fe53ee1e8850d544f7690bc32