Switched video posting from feathers client request to axios post.
The feathers client is connecting to the API server. This is fine for
most things, but video uploads need to go to the media server.
The quick solution was to make createVideo post an HTTP request to
/video so that the media server handles it.

Removed unneeded <EmptyLayouts> from admin console components since
they come with alert divs that we don't want.

Switched order of description and url inputs in video modal.


Former-commit-id: 7166c4f060fbbc5273f5d07cd5654e4c707f4e15