npm 'test' script now uses compiled codebase.
Due to weirdness with running jest on the Next server with
NODE_ENV=development, 'npm run test' now runs the compiled tests
in the lib directory.

Updated app.test.ts to listen for the 'next-ready' event if
running in development mode but to just wait for 'server:listening'
if not, as in production mode the next server is already compiled.


Former-commit-id: 4fa200fb81e36a0c2d8110ad12f3ced2c169b45f