npm 'test' script now uses compiled codebase.
Due to weirdness with running jest on the Next server with
NODE_ENV=development, 'npm run test' now runs the compiled tests
in the lib directory.

Updated app.test.ts to listen for the 'next-ready' event if
running in development mode but to just wait for 'server:listening'
if not, as in production mode the next server is already compiled.


Former-commit-id: d780d183a00b07761ad9fba4bb8f1d42fe89df58