Fixed a few bugs in Spoke custom API.
Spoke API had an incorrect reference to the data returned from the
XR3ngine API. It also wasn't signing requests in fetch().

The relationship between static_resource and attribution had been
reveresed but the add-attribution hook hadn't been updated to
reflect that.

Updated user test so that beforeAll always makes a User for the
Update and Delete tests to act on. They were occasionally not
getting any results.


Former-commit-id: 050d30e38998ea7ee47666c4990cf68391da5b35