Revamped login and invite system.
Login now uses a bearer token system instead of sending the full JWT
in the link. This allows for SMS logins that will fit in a single text
message, which a full JWT would not. Bearer tokens are exchanged for a
JWT and the user is redirected to the existing front-end magiclink
route.

SMS invite templates are now under 160 characters.

Invalid JWT logins now redirect to '/'; previously, they were going to
'#', which would leave users stuck on the 'waiting' page indefinitely.

_app.tsx checks if there's an 'error' query parameter. If so, it
dispatches that error and removes that query parameter. This allows the
client to handle non-critical errors like failed login attempts.

Fixed bugs in SMS invites.
* Client was not appending '+1' to front of token
* Incorrect template path
* Service title was not being properly provided to templates.

Changed accept-invite route and service name to 'a-i'. This was only
done so that links were shorter, but feathers requires path and service
name to match.

Fixed some bugs in friend invite system.

Made JWTs expire in 30 days instead of default 1 day.


Former-commit-id: 7a892f94f959bbac6f021e536e0eeed5a4642126