Rewrote some of the stream rendering to try to better use observables. At this point, I think the Consumers may not be fully observable, as I can't seem to have anything triggered on the pause state changing.
Producers and consumers are now pausable. Pausing video will turn the <video> element to display: none, and show the
profile outline background. Pausing a video consumer will do so for all consumers of that video.

If a producer is paused, all consumers' pause/resume buttons for it will disappear to indicate that the consumer user
can't control it, and to prevent them from 'unpausing' and nothing appearing to happen.


Former-commit-id: 032b92412afac371aa63215b42ba0ad901ab423e [formerly 1dd92ec7825352a69d5f6a5714b49c8b319768db]
Former-commit-id: f2eadbb7cba446ea60f28eb68ebb4173c49e0e08