Accepting party invites placed user on the party's instance server, if there is one.
Started implementing logic to check if adding a party member exceeds the location's maxUsers
and, if so, moves the party to an instance that has enough room or spins up a new one.
Running into a race condition where client-side provisioning of old instance server
is occurring after provisioning of new server.


Former-commit-id: 2e04088ba4271cef1dab438371fee0a01470092c [formerly 29ce83d5359f3f4a33834e637d2a73ba49797bb1]
Former-commit-id: c002c43cb3ddced88a95221e5c1e9578f84a77e0