Updated oauth flow.
Client now calls /oauth/<service> with the JWT, which
feathers-oauth will use to upgrade the current user rather
than making a new one.

Added some overrides to oauth strategy updateEntity call
to make userRole 'user' and delete current guest user if
it's not being upgraded (i.e. if a user associated with
that oauth account already exists, and we're signing in
with them rather than upgrading a guest account).


Former-commit-id: 81d961e5041c02703587bcba658b199107fd3ef4 [formerly 713a3288a053d9f666012113f999d9bf4412b626]
Former-commit-id: 0058510cc822444c1b7af9cb4d6c7919092f1b56