Added some cleanup code to close leftover transports.
Sometimes, the client signals to the gameserver that it should close DTLS.
This was causing client disconnects to not trigger the normal transport
cleanup, which was leading to some weird behavior when that user
reconnected, namely video and audio streams that were muted.
There's now a call to closeTransport when dtlsState changes to 'closed'
and a call to clean up any transports that match peer, direction, and party
when a new transport is created.


Former-commit-id: 613250066950490de9e21a277b65269158b60119 [formerly 1b72b0c4953abb7f15fc89a92d28ce3dd19d6ad6]
Former-commit-id: 7d3dada6d511d597b17235f46a8355bf4ca049c6