Hackfix to model upload flow.
There was a seemingly-unnecessary POST to /static-resource in _uploadAsset in Editor/Api.tsx
that was erroring out due to not sending the correct fields for a static-resource.
It appears that the static-resource is already created in a prior POST to /media.
The call to /static-resource has been commented out and the return object from _uploadAsset
has been slightly modified to use what's returned from .upload.

This should be looked at further, though, since I'm guessing what was commented out is
supposed to do something useful.


Former-commit-id: fce8ed20afde20e77cca6e60fd610e58312f6e84 [formerly c5d774967a46c9e0428a778f6aff6a8b9cdb62fe]
Former-commit-id: 35944a9621e08539d7996126c8f6fca86200587b