Hackfix to model upload flow.
There was a seemingly-unnecessary POST to /static-resource in _uploadAsset in Editor/Api.tsx
that was erroring out due to not sending the correct fields for a static-resource.
It appears that the static-resource is already created in a prior POST to /media.
The call to /static-resource has been commented out and the return object from _uploadAsset
has been slightly modified to use what's returned from .upload.

This should be looked at further, though, since I'm guessing what was commented out is
supposed to do something useful.


Former-commit-id: c5d774967a46c9e0428a778f6aff6a8b9cdb62fe