Added subdomain information to instances in admin panel.
Made instance modal close if there are no more users on it.

Added some user-is-admin checks to backend services.
Updated handful of other calls to vanilla user.find()
to use 'action: layer-users' so as not to be blocked
by new admin check. All of these calls seemed to
be getting users on the current layer/instance anyway.

Fixed some styling issues.


Former-commit-id: 1cce07050e9c97e5a62941ea9a1c0c609058e907 [formerly 546068373f7fb5f1b045eb997bc36f780332aae6]
Former-commit-id: c4fa9e184c6b81e1f57b5a7cb772d11a6011e0eb