Fixed some build and deployment issues.
DRACOLoader.js and GLTFLoader.js were changed from 'export' to 'module.exports',
and for some reason this caused imports of them to not recognize what
was being imported. Reverted back to 'export'.

The check for whether a DB is running locally or as a Docker container should not
be run in production, as the DB is run externally. Added conditional statement
to skip that check in production.


Former-commit-id: 3727704583433555d213a8173d1b12b84cdafbd9 [formerly 81f961a793ed7f0719160cd9a0005fd2f8bc3002]
Former-commit-id: c8502d437c66bd9ae25bceec9b141f2a46e2e70e