Fixed bug in logout.
Turned out to just be an issue with a non-conditional reference to authUser.

Started fixing behavior when user logs out while in a location. Old user is removed,
new user is added to world, but scene is not being reloaded. Also, for some reason
the old network ID is being sent, possibly along with the new one.

Need to find a better way to determine that user is connected to world before
calling JoinWorld other than just setTimeout(3000).


Former-commit-id: 0a80e129878a7ce12cdb5e008905f3216d267bc2