Fixed some bugs with Harmony rendering and connecting.
On some occasions, group/friend/party chat was not connecting a second user to
the first's producers. This ended up being due to how Mediasoup routers were being created.
An instance router is made at GS startup, but GS code was waiting until a transport was
being created to make a router for a channel. If requests to make send and receive transports
came it at basically the same time, both were resulting in router creation, but one was getting
overwritten by the other, leading future consumers to be trying to consumer a different router,
which can't happen.

Channel router is now created when a GS is being initialized as a channel instance server, well before
transport creation. Creating router if it doesn't exist on transport creation was left as-is as a fallback.

Fixed bug where initial channel self-user window wasn't rendering as a 'single' since layerUsers.length
was undefined instead of a number.

Removed a lot of seemingly-unnecessary console.logs.


Former-commit-id: db3a345b0d4b2d1d1f87899a4c906740399348d9 [formerly ef2508965f94ec017661bbf55e66bc39848e688c]
Former-commit-id: 825700d18e975e7f3d662d4f975875430d35f9db