Untangled circular dependency between client and client-core with redux.
client was assembling reducers from client-core, but then client-core store.ts was importing
reducers from client. Updated store.ts to export a class with a static 'store' that configureStore()
can remake with reducers that are passed as a parameter: `configureStore(reducers)`.

social was likely using client's reducers, and is now using its own.

Updated client and social to follow this new implementation.


Former-commit-id: adce9b9d16bd7d4f3267c16b7f1e6f05838856cf