Added ability to pre-load location on gameserver.
For events, it is useful to have a number of gameservers already set to a location, as loading
complex scenes can take an appreciable amount of time.

If ENV_VAR PRELOAD_LOCATION_NAME is set on a gameserver, it will automatically load that location
and allocate itself with 0 users on startup. New users for that location will be provisioned to
one of the preloaded servers.


Former-commit-id: c9ae71504500d9421127fd4a99e28ca7b419a25f