Fixed several bugs and improved content-packs.
Updated process of adding avatars and scenes to content packs so
that multiple ones can be added at once.

Made it so that the first non-guest user to sign in is an admin.

Fixed a bug in add-file-id where a null value for ownedFileIds was failing
due to trying to read a value of null.

Downloading scenes was using the configured S3 bucket for component
URLs instead of the URL of the content-pack. This has been fixed by getting the
URL from the manifest URL.

Updated doLoginAuto so that forceClientAuthReset will skip using the old accessToken,
thus making logout actually work.

Changed a bunch of 'onChangePage' and 'onChangeRowsPerPage' to 'onPageChange' and
'onRowsPerPageChange' since warnings were being thrown about the former being deprecated
for the latter.


Former-commit-id: c4f0b996007eaf72fd63a21a0f805ad7799fc94e