Fixed a bug with feathers channels and separated client media streams.
In server/src/channels.ts, the identity-provider was not always showing up on the connection input.
This was causing users to not subscribe to their own feathers channel, and weren't getting any updates.
Since the identity-provider was appearing on the authResult, made identityProvider sourced from either
input.

Testing uncovered that mic-only devices couldn't start their media streams since the codebase was always
requesting both audio and video, and the MediaDevices API will reject a request if one requested device
is not present. In MediaStreamSystem.ts, made separate videoStream and audioStream properties to replace
mediaStream property. Associated functionality now only deals with the device/stream that it needs, i.e.
audio functionality deals only with audioStream. Semi-duplicated some functions so that there are separate
handlers for audio and video.

Updated A/V control buttons so that each is only displayed if there's a device to capture that stream.


Former-commit-id: 57b79b63984c75c76a612b6674f8c2264d4ae7c7