Fixed bug with handling Authorization message more than once.
Sending Authorization message at an interval rate was causing multiple socket
listeners to be created if the first call wasn't fully handled within the interval.
SocketWebRTCServerTransport.ts now tracks if the listeners have been created so
that it only makes them once.

Fixed a few bugs with GameServerWarnings.tsx. The noCountdown option should always be
set to override any other case of it being set (though maybe the real issue is that
the WebGL disconnected error is triggering when the real issue is something else,
like a location not existing or a gameserver going down).


Former-commit-id: 4fcfcc56e2be738e0670a6d5ba019ec1437a08ea