MAP-45: create task and phase abstractions
Needed greater flexiblity and transparency with how the various tasks
involved in producing the map dispatched. When simply using a call tree
i.e. functions that defer to other functions, the code was
difficult to reason about, especially when it came to when it came to
timing. Does the code wait for all tasks of a given phase (e.g. fetching
tiles) to complete before moving on to the next? Or does it begin the
next phase concurrently as soon as some of the tasks of the earlier
phase have completed? These are the kind of questions I hope I've made
obvious.


Former-commit-id: e070aa4a325a4967f0b12c42b90cf5084dec9635