Implemented builder service for in-K8s building.
The client needs to be rebuilt whenever content-packs are updated.
The build process is now as follows:

Github Actions runs npm install on the new commit and copies all files
into the Docker environment, but does not run build-docker. This image
is deployed to a builder service within the cluster. The builder service
downloads content packs, then builds the docker image with those content
packs and pushes/deploys them. The service sleeps infinitely once it's
done so that the pod doesn't restart on its own.

The final built image is tagged with the git commit hash and the datetime.
This is so that rebuilds of a particular deployment due to new reality packs
are uniquely identified from each other.


Former-commit-id: d8daee2bba88ba0d645284abcbf4c59145a4f9de